Improve gcov support
If someone passing CXX value to configure that is not in their
path then they will also need to pass in matching GCOV if they
want gcovr target to work.  Add support for this.

Also, minor improvements to fail if gcov is not found and print
info on needing --enable-gcov when running gcovr target.